billiard: Adjust scale of billiard table to real scale
We have ragged physics simulation in that a billiard ball's radius is
measured exactly 1.  We must use metric gauge in our simulation in order
to cope with Oculus Rift, but the simulation exposed its improper
implementation in metric.  We want to debug its behavior in more easy way
than Rift (we don't want to put on Rift and walk around for each debug
iteration of physics simulation).

Just making it to scale was a tough job since all the codes have hard-
coded 1 as the balls' radius, but it's done.  The project's behavior is
temporarily broken, but we have to fix it in order to proceed anyway.

OculusBilliard is not touched yet.