billiard: Properly delete the context of WGL on exit
There's a strange issue that explorer and other applications' graphics
corrupt when OpenGL application is repeatedly created and destroyed.
This issue is fairly rare, we might see it about once two weeks without
restarting the OS.  Once this happened, there are no way but restarting
the computer to fix the state.

This phenomenon is observed for other OpenGL games and the reason is not
known, but it seems to hang around since I have had Windows 7.
This specific OpenGL application (billiard) omits WGL context cleanup
on program termination.  Doing this properly may or may not help.