billiard: Delete project file for old Visual Studio
The project file and the solution file are placed in a directory named
Monster, which was name of a computer.  This naming convention does not
cope with open source distributedly version controlled project, hence
deleted.

Now that we have VC2012 project files in another directory, we really
do not need older versions of the project files.