obilliard: Enable the 2nd joystick to control head rotation
It works as a replacement to mouse yaw control.  When the player holds a
gamepad, he cannot easily switch hand to mouse, since his vision is
blocked by Rift :)  This little change helps greatly in controlling the
player in the virtual space (provided that he has a gamepad for PC).